moved statsChart to DataDisplay folder
