+++ src/components/index.ts
@@ -12,8 +12,8 @@ export * from './buttons/BackToButton'
 export { default as UploadButton } from './buttons/UploadButton'
 export * from './buttons/UploadButton'
 
-export { default as StatsChart } from './charts/StatsChart'
-export * from './charts/StatsChart'
+export { default as StatsChart } from './dataDisplay/StatsChart'
+export * from './dataDisplay/StatsChart'
 
 export { default as DeprecatedStatsChart } from './charts/DeprecatedStatsChart'
 export * from './charts/DeprecatedStatsChart'
